+++ sdk/cosmos/azure-cosmos/azure/cosmos/aio/_global_endpoint_manager_async.py
@@ -25,9 +25,10 @@ database service.
 
 import asyncio # pylint: disable=do-not-import-asyncio
 import logging
-from typing import Tuple, Any
+import os
+from typing import Any
 
-from azure.core.exceptions import AzureError
+from azure.core.exceptions import AzureError, ServiceRequestError, ServiceResponseError
 from azure.cosmos import DatabaseAccount
 
 from .. import _constants as constants
@@ -50,7 +51,6 @@ class _GlobalEndpointManager(object): # pylint: disable=too-many-instance-attrib
         self.client = client
         self.PreferredLocations = client.connection_policy.PreferredLocations
         self.DefaultEndpoint = client.url_connection
-        self.refresh_time_interval_in_ms = self.get_refresh_time_interval_in_ms_stub()
         self.location_cache = LocationCache(
             self.DefaultEndpoint,
             client.connection_policy
@@ -61,9 +61,6 @@ class _GlobalEndpointManager(object): # pylint: disable=too-many-instance-attrib
         self.refresh_lock = asyncio.Lock()
         self.last_refresh_time = 0
         self._database_account_cache = None
-
-    def get_refresh_time_interval_in_ms_stub(self):
-        return constants._Constants.DefaultEndpointsRefreshTime
 
     def get_write_endpoint(self):
         return self.location_cache.get_write_regional_routing_context()
@@ -115,8 +112,11 @@ class _GlobalEndpointManager(object): # pylint: disable=too-many-instance-attrib
                 await self.refresh_task
                 self.refresh_task = None
             except (Exception, asyncio.CancelledError) as exception: #pylint: disable=broad-exception-caught
-                logger.exception("Health check task failed: %s", exception) #pylint: disable=do-not-use-logging-exception
-        if current_time_millis() - self.last_refresh_time > self.refresh_time_interval_in_ms:
+                logger.error("Health check task failed: %s", exception, exc_info=True)
+        if current_time_millis() - self.last_refresh_time > int(os.getenv(
+                constants._Constants.AZURE_COSMOS_DATABASE_ACCOUNT_REFRESH_INTERVAL_IN_MS,
+                str(constants._Constants.DefaultEndpointsRefreshTime)
+        )):
             self.refresh_needed = True
         if self.refresh_needed:
             async with self.refresh_lock:
@@ -138,40 +138,45 @@ class _GlobalEndpointManager(object): # pylint: disable=too-many-instance-attrib
                 self.refresh_needed = False
                 self.last_refresh_time = current_time_millis()
                 if not self.startup:
-                    # this will perform getDatabaseAccount calls to check endpoint health
+                    # this will perform both database account and checks for endpoint health
+                    # in background
+                    self.refresh_task = asyncio.create_task(self._refresh_database_account_and_health())
+                else:
+                    database_account = await self._GetDatabaseAccount(**kwargs)
+                    self.location_cache.perform_on_database_account_read(database_account)
+                    # this will perform only calls to check endpoint health
                     # in background
                     self.refresh_task = asyncio.create_task(self._endpoints_health_check(**kwargs))
-                else:
-                    # on startup do this in foreground
-                    await self._endpoints_health_check(**kwargs)
                     self.startup = False
 
-    async def _database_account_check(self, endpoint: str, **kwargs: dict[str, Any]):
+    async def _refresh_database_account_and_health(self, **kwargs):
+        database_account = await self._GetDatabaseAccount(**kwargs)
+        self.location_cache.perform_on_database_account_read(database_account)
+        await self._endpoints_health_check(**kwargs)
+
+    async def _health_check(self, endpoint: str, **kwargs: dict[str, Any]):
         try:
-            await self.client._GetDatabaseAccountCheck(endpoint, **kwargs)
+            await self.client.health_check(endpoint, **kwargs)
             self.location_cache.mark_endpoint_available(endpoint)
-        except (exceptions.CosmosHttpResponseError, AzureError):
-            self._mark_endpoint_unavailable(endpoint)
+        except (exceptions.CosmosHttpResponseError, AzureError) as exception:
+            if isinstance(exception, (ServiceRequestError, ServiceResponseError)):
+                self._mark_endpoint_unavailable(endpoint)
 
     async def _endpoints_health_check(self, **kwargs):
         """Gets the database account for each endpoint.
 
         Validating if the endpoint is healthy else marking it as unavailable.
         """
-        # get the database account from the default endpoint first
-        database_account, attempted_endpoint = await self._GetDatabaseAccount(**kwargs)
-        self.location_cache.perform_on_database_account_read(database_account)
         # get all the endpoints to check
         endpoints = self.location_cache.endpoints_to_health_check()
-        database_account_checks = []
+        health_checks = []
         for endpoint in endpoints:
-            if endpoint != attempted_endpoint:
-                database_account_checks.append(self._database_account_check(endpoint, **kwargs))
-        await asyncio.gather(*database_account_checks)
+            health_checks.append(self._health_check(endpoint, **kwargs))
+        await asyncio.gather(*health_checks)
 
         self.location_cache.update_location_cache()
 
-    async def _GetDatabaseAccount(self, **kwargs) -> Tuple[DatabaseAccount, str]:
+    async def _GetDatabaseAccount(self, **kwargs) -> DatabaseAccount:
         """Gets the database account.
 
         First tries by using the default endpoint, and if that doesn't work,
@@ -179,12 +184,12 @@ class _GlobalEndpointManager(object): # pylint: disable=too-many-instance-attrib
         specified, to get the database account.
         :returns: A `DatabaseAccount` instance representing the Cosmos DB Database Account
         and the endpoint that was used for the request.
-        :rtype: tuple of (~azure.cosmos.DatabaseAccount, str)
+        :rtype: ~azure.cosmos.DatabaseAccount
         """
         try:
             database_account = await self._GetDatabaseAccountStub(self.DefaultEndpoint, **kwargs)
             self._database_account_cache = database_account
-            return database_account, self.DefaultEndpoint
+            return database_account
         # If for any reason(non-globaldb related), we are not able to get the database
         # account from the above call to GetDatabaseAccount, we would try to get this
         # information from any of the preferred locations that the user might have
@@ -197,10 +202,9 @@ class _GlobalEndpointManager(object): # pylint: disable=too-many-instance-attrib
                 try:
                     database_account = await self._GetDatabaseAccountStub(locational_endpoint, **kwargs)
                     self._database_account_cache = database_account
-                    self.location_cache.mark_endpoint_available(locational_endpoint)
-                    return database_account, locational_endpoint
+                    return database_account
                 except (exceptions.CosmosHttpResponseError, AzureError):
-                    self._mark_endpoint_unavailable(locational_endpoint)
+                    pass
             raise
 
     async def _GetDatabaseAccountStub(self, endpoint, **kwargs):