+++ sdk/cosmos/azure-cosmos/tests/test_crud.py
@@ -15,7 +15,7 @@ import os
 import pytest
 import requests
 from azure.core import MatchConditions
-from azure.core.exceptions import AzureError, ServiceResponseError
+from azure.core.exceptions import AzureError, ServiceResponseError, ServiceRequestError
 from azure.core.pipeline.transport import RequestsTransport, RequestsTransportResponse
 from urllib3.util.retry import Retry
 
@@ -1166,21 +1166,20 @@ class TestCRUDOperations(unittest.TestCase):
 
             # client does a getDatabaseAccount on initialization, which will not time out because
             # there is a forced timeout for those calls
-            client = cosmos_client.CosmosClient(TestCRUDOperations.host, TestCRUDOperations.masterKey, "Session",
+            with self.assertRaises(ServiceRequestError):
+                client = cosmos_client.CosmosClient(TestCRUDOperations.host, TestCRUDOperations.masterKey, "Session",
                                                 connection_policy=connection_policy)
-            with self.assertRaises(Exception):
                 databaseForTest = client.get_database_client(self.configs.TEST_DATABASE_ID)
                 container = databaseForTest.get_container_client(self.configs.TEST_SINGLE_PARTITION_CONTAINER_ID)
                 container.create_item(body={'id': str(uuid.uuid4()), 'name': 'sample'})
 
-    def test_read_timeout_async(self):
+    def test_read_timeout(self):
         connection_policy = documents.ConnectionPolicy()
         # making timeout 0 ms to make sure it will throw
         connection_policy.DBAReadTimeout = 0.000000000001
-        with self.assertRaises(ServiceResponseError):
-            # this will make a get database account call
-            with cosmos_client.CosmosClient(self.host, self.masterKey, connection_policy=connection_policy):
-                print('initialization')
+        # this will make a get database account call
+        with cosmos_client.CosmosClient(self.host, self.masterKey, connection_policy=connection_policy):
+            print('initialization')
 
     def test_client_request_timeout_when_connection_retry_configuration_specified(self):
         connection_policy = documents.ConnectionPolicy()
@@ -1195,11 +1194,11 @@ class TestCRUDOperations(unittest.TestCase):
         )
         # client does a getDatabaseAccount on initialization, which will not time out because
         # there is a forced timeout for those calls
-        with cosmos_client.CosmosClient(self.host, self.masterKey, connection_policy=connection_policy) as client:
-            with self.assertRaises(AzureError):
-                databaseForTest = client.get_database_client(self.configs.TEST_DATABASE_ID)
-                container = databaseForTest.get_container_client(self.configs.TEST_SINGLE_PARTITION_CONTAINER_ID)
-                container.create_item(body={'id': str(uuid.uuid4()), 'name': 'sample'})
+        with self.assertRaises(ServiceRequestError):
+            with cosmos_client.CosmosClient(self.host, self.masterKey, connection_policy=connection_policy) as client:
+                    databaseForTest = client.get_database_client(self.configs.TEST_DATABASE_ID)
+                    container = databaseForTest.get_container_client(self.configs.TEST_SINGLE_PARTITION_CONTAINER_ID)
+                    container.create_item(body={'id': str(uuid.uuid4()), 'name': 'sample'})
 
     # TODO: Skipping this test to debug later
     @unittest.skip