+++ sdk/cosmos/azure-cosmos/tests/test_crud_response_payload_on_write_disabled.py
@@ -16,9 +16,8 @@ import uuid
 import pytest
 import requests
 from azure.core import MatchConditions
-from azure.core.exceptions import AzureError, ServiceResponseError
+from azure.core.exceptions import ServiceRequestError
 from azure.core.pipeline.transport import RequestsTransport, RequestsTransportResponse
-from urllib3.util.retry import Retry
 from typing import Any, Dict, Optional
 
 import azure.cosmos._base as base
@@ -2031,9 +2030,9 @@ class TestCRUDOperationsResponsePayloadOnWriteDisabled(unittest.TestCase):
             connection_policy.RequestTimeout = 0.000000000001
             # client does a getDatabaseAccount on initialization, which will not time out because
             # there is a forced timeout for those calls
-            client = cosmos_client.CosmosClient(self.host, self.masterKey, "Session",
-                                                connection_policy=connection_policy)
-            with self.assertRaises(Exception):
+            with self.assertRaises(ServiceRequestError):
+                client = cosmos_client.CosmosClient(self.host, self.masterKey, "Session",
+                                                    connection_policy=connection_policy)
                 databaseForTest = client.get_database_client(self.configs.TEST_DATABASE_ID)
                 container = databaseForTest.get_container_client(self.configs.TEST_SINGLE_PARTITION_CONTAINER_ID)
                 container.create_item(body={'id': str(uuid.uuid4()), 'name': 'sample'})